requirements.txt
# Django
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4

# Network & Utils
requests==2.23.0

# Cryptography & Security
six==1.15.0
pyjwt==1.7.1
jwcrypto==0.7